scale_fill_radiohead_d Uses the radiohead album selected to create a discrete fill

scale_fill_radiohead_d Uses the radiohead album selected to create a discrete fill

scale_fill_radiohead_d(
  ...,
  alpha = 1,
  begin = 0,
  end = 1,
  direction = 1,
  album = "in_rainbows",
  values = NULL,
  space = "Lab",
  na.value = "grey50",
  guide = "colourbar",
  aesthetics = "fill",
  discrete = TRUE
)

Arguments

alpha
pass through parameter to ggplot

begin
The (corrected) hue in 0,1 at which the radiohead colormap begins.

end
The (corrected) hue in 0,1 at which the radiohead colormap begins.

direction
Sets the order of colors in the scale. If 1, the default, colors are as output by viridis_pal. If -1, the order of colors is reversed.

album
the radiohead album to select, e.g. album = "ok_computer"

discrete

Value